Moved StartUp functionality to Program
